requirements.txt
asgiref==3.5.2
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.2
whitenoise==6.2.0
